Ostensibly functional row_dists, plus wrappers
row_dists seems to work with both Jensen-Shannon and correlations, so
that's good. The topic_divergences() wrapper seems to choke on the wide
dataframe. Fixed incorrect topic_words_wide() to produce a topic x word
dataframe with dcast. That seems fine...

Also inserted on this branch, because I was working on this when I
noticed the need: utility functions for normalizing and smoothing raw
counts. Not otherwised used right now.