matrix utils incl. time series
Rewriting the matrix utilities to be more generic and compositional,
with just a couple of time-series shortcuts. This replaces the hopeless
tangle of topic_yearly this and term_journal_topic that. The basic
munging issue of having to switch from matrix to data frame somewhere
along the line remains, but is at least consolidated by gather_matrix.