rmultinom_sparse bug fix
shouldn't allocate anything to zero-probability words, so we have to
ensure the category for the remainder n_left at the end of the loop in
sparse_multinomial_trial is the last non-zero probability category. We
do this by truncating the probability vector at its last non-zero
element, which should save us a little time too.