cleanup first draft a little
still untested, but instead of a bunch of arrays this rewrites and sorts
D as a vector of little `pair_dist` structs, which is a little clearer.
We also avoid using a lambda and thus avoid needing the c++11 plugin.

Sign of the threshold check was backwards: Stop when our distance
is GREATER than the threshold. Duh.