lowered expectations
when we merge two clusters, the result always gets the lower number.

I thought this would fix the case in which threshold > max(dist) but not
all clusters were full. But it seems that's not so: we can lose and end
up with unmerge-able non-maximal clusters. We can, however, still expect
to get no isolates in this case. So I changed the test to test for this.