fix: scrambling of D in naive_cluster
as is traditional, sharing code immediately provokes realization of
obvious bug. The ordering of D as assumed in naive_cluster did not
correspond to the ordering supplied from align_topics (or
unnest_model_distances). Which should have been obvious from the way my
comment in the former did not correspond to the way model_distances was
written. Anyway, this is now fixed by the application of yet more brute
force.

Added two more tests to try to catch stupidities like this by their
results on really easy clustering problems. Clustering a model with
copies of itself should go really well and in the buggy code it didn't.
With the fix this test seems to pass and the results from test models are more
plausible-seeming in general.