change model_distances to take matrix-by-matrix g
This changes the expectations about the `g` argument of model_distances.
Instead of operating on pairs of vectors, it is expected to operate
on pairs of matrices. It was silly to force everything through a
double-apply bottleneck. Premature optimization is the root of all etc.
but I just couldn't stand it. With the updated JS_divergence this can
still be used as the default. ?model_distances explains how to turn a
distance on vectors into one on matrix rows using apply.

Also get rid of one layer of batty transposing: a model_distances now,
more sensibly, should have D(m1:k1, m2:k2) in row k1, column k2. Updated
the subscript and unnest functions correspondingly, and the
tests that checked these specifications.