data-structure tweaks; cluster_widths -> widths
for a little more user convenience, we store the parameters to
model_distances on the dst object, and the parameters to align_topics
on the topic_alignment object. (Couldn't get a nice unevaluated "call"
to do what I want). That way you can just send the single argument to
alignment_frame and widths (formerly known as cluster_widths).
Documented accordingly. Tests adjusted as well.

My widths check also needed to deal with the case where a topic is an
isolate, in which case the "width" of its cluster is zero.