1st step for aligning models with varying K
Which is do away with the assumption that K is the same for each model.
Do this by just wasting even more space storing indices in
naive_cluster. We pass in a vector of model sizes so that we can count
along the unrolled D correctly, storing model and topic numbers for each
entry now. align_topics now returns cluster numbers not in a matrix but
in a list since this may be ragged. Various changes necessary to
accommodate this data structure change.

Including multiple adjustments to tests. These now pass, but I am not
sure I am getting the same clusterings as before. The results look
somewhat worse than they did, but maybe that's just today's random
number generation being less favorable than yesterday? Not catching any
obvious errors so far.