threshold test correction
we do d->d > threshold in naive_cluster, so we have to expect d <=
threshold in the tests.

Was getting sporadic failures when there were clusters exactly at the
threshold! But that was because the test was wrong: we are closed, not
open, on the boundary.

Also, since we're using randomization in generating the models, added a
message so we can reproduce any particular test.