Abstract away differences between mallet 1.0/1.1
The key is to use ParallelTopicModel where we want a P.T.M. method, and
then rewrite ParallelTopicModel to check whether the mallet_model$model
object is an instance of PTM or has a PTM data member.

The other thing we need to do is to deal with mallet 1.0's assumption
that rJava is in its search path. When we load_mallet(), if we detect
mallet 1.0, we explicitly `require("rJava")`. Otherwise we can follow
the rules a little more and skip that step.

Tests seem to pass after installing either mallet version, though it's a
little hard to keep track.