remove unreasonable dense-cluster-seq. expectation
This was silly, but it took many runs to discover it with my tests (bad
tests). Random seed 4189 revealed it: if you have three models, it is
perfectly possible to terminate with a sequence of cluster ID's with
gaps if you have a state where the remaining non-full clusters are

1.a 2.b
1.c 3.d
2.e 3.f

None of these can be merged, and 2.e 3.f will not have a cluster number
in sequence (its number is the sequential number of 2.e, i.e. K1 + e).