model_distances m/Matrix conversion
matrix and Matrix do not play well together. Rather than leave the job
of worrying about this entirely to the package user who wishes to vary
`g`, model_distances checks if there is a type disparity and casts both
matrices down to `matrix` if so. Added a test for this.