fix bad info export: auto_unbox=TRUE
no auto_unbox of default `info` was generating a bad info.json file (or
m__DATA__info JSON string ) with arrays in place of single values. This
caused an error when dfr-browser tried to `utils.deep_replace` its
default VIS with `VIS` from info.json...but only in WebKit! I guess
Firefox was more tolerant.

My R tests missed this because of course in R a single value IS a
one-element vector. Sigh.