option to export normalized matrices for browser
just in case someone wants to see how much difference using the sampling
weights instead of posterior probabilities makes. The key API change is
the addition of `proper` and `digits` parameters to
`export_browser_data`. If `proper` is true, we smooth and normalize,
then round to `digits`, sparsify, and export. The default behavior
remains as before.

To implement this a little more neatly, I factored out the individual
data file exports from `export_browser_data`. These can now be
invoked separately as well.

Added a test of the "proper" export, which seems to work plausibly.