correct an alignment test
sigh. in aligning models from three packages, we can't just check that
the max cluster number is small, since we can get gaps in numbering even
if the alignment algorithm is behaving correctly. Instead we check that
we have no isolates, which really should be impossible for three models
with 8 topics each and no ceiling on linkage distance.