workaround bigtabulate glitch
the latest bigtabulate (1.1.5) was leading to errors in
`load_from_mallet_model` which invokes `bigtabulate::bigsplit`. After
some guesswork I believe the error is indeed a bug in `bigtabulate` but
can be worked around by passing `splitcol=NA_real_` in place of the
default `splitcol=NA` option to `bigsplit`. Updating the package so it
remains usable until this issue is addressed in less kludgy fashion.