read_state_nobigmem: use simplified state
We are still simplifying the state even if we are loading it as a data
frame and not as a big.matrix, so why not load that instead of the gibbs
state? (the gibbs-reading code is still useful; it can be used
for an all-R simplify_state. See next commit.)