simplify_state in R (no python)
this repurposes the gibbs-state-reading code I wrote for
read_state_nobigmem. simplify_state reads the MALLET state in chunks and
writes out the simplified state in chunks. The chunk size is a new
package option, dfrtopics.state_chunk_size.

That's probably slower than what python can do, but having a python
dependency just for this one operation wasn't great. Learned that one
package user had to do a lot of work to satisfy the dependency on a
system where python 3 was the default python (and simplify_state.py is
not python3-happy).

Reading gz files in chunks is a little tweaky and I'm not totally
convinced what I have here is robust. Trying to read in chunks from a
gzfile stream doesn't work, but reading in chunks from a gzcon(file())
does. However readr::read_delim doesn't seem to like this, so we fall
back on ordinary read.table.