fix issue with comment.char in read_gibbs
we can't allow "#" as a comment character for read.table because tokens
might have "#" in them (sigh). Instead we just read past the header 3
lines of comments in simplify_state (and issue a warning if it doesn't
look right).

Did a little benchmarking on whether it might be better to parse the
lines from readLines using str_split_fixed myself in read_gibbs. No:
surprisingly, read.table better choice (presumably because we don't
allocate and delete the space for the fields we discard).

Still 1/2 as fast as the python script, though. Sigh. Added comment with
speed estimate.