picking up again on dfb multimodel export
Because I am a bad programmer, I dropped this without carefully noting
just where things were. I appear to have drafted all the export code and
written documentation but not tried to test things.

Also, in the interim roxygen2 and Rcpp got fiddled around so all the
docs and Rcpp exports are generated in this commit