dfr_browser with multimodel & alignment working
another pass at dfr_browser.list, cleaning things up. Went back to
metadata and bib type "dfr" as default because that's probably less
unexpected, and the "base" types are kind of obscure. This necessitates
some special-casing to deal with the truly dumb fact that dfr-browser
refers to the "pubdate" field as "date" and the "journaltitle" field as
"journal."

dfr_browser.topic_alignment looks like it works too. Lots of fiddling
with the way the various info.json files are prepared needed to get this
going, probably lots of hidden bugs hanging around.